backend/pyproject.toml
[tool.poetry]
name = "django-skeleton"
version = "0.1.0"
description = "Django skeleton with most useful functionalities for new projects."
authors = [""]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
Pillow = "^8.3.1"
sentry-sdk = "^1.3.1"
celery = "^5.1.2"
redis = "^3.5.3"
django-allauth = "^0.45.0"
django-rest-auth = "^0.9.5"
django-extensions = "^3.1.3"
django-anymail = "^8.4"
django-phonenumber-field = "^5.2.0"
kombu = "^5.1.0"
raven = "^6.10.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.1"
drf-yasg = "^1.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


#[tool.poetry.scripts]
#migrate = "python manage.py migrate"
#makemigrations = "python manage.py makemigrations"
#server = "python manage.py runserver"
#createsuperuser = "python manage.py createsuperuser"
#shell = "python manage.py shell_plus"
#show_urls = "python manage.py show_urls"
#celery = "celery -A config.celery worker -B -l info"
#collectstatic = "python manage.py collectstatic --no-input"